muebles_lottus_backend/requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
